Added CSS scope prefix for OL3/5 sidebar.
When used in conjunction with other CSS frameworks, the more generic
names of some of the CSS classes can easily result in conflicts. As all
CSS classes of the main OpenLayers project use the 'ol-' prefix, it
seems reasonable to follow this convention also for the sidebar. As the
OL2 version uses the generic JS, the prefixes cannot easily be applied
here.